Fixes sounds missing, when client needs data from hotfix.

+++ Source/Game/DataStorage/ClientReader/DB6Storage.cs
@@ -305,7 +305,7 @@ namespace Game.DataStorage
                                     locale = 0;
                                     if (!locStr.HasString(locale))
                                     {
-                                        buffer.WriteUInt16(0);
+                                        buffer.WriteUInt8(0);
                                         break;
                                     }
                                 }
